Moved game logic to its own class from mainwindow
